Updated Wallpaper selection process
1. A problem was found with the old wallpaper selection command, so I have updated the entire process. The old way involved an OR statement where after the first statement becomes true, the second will never be true again except under rare conditions. 

The process now checks to see if there are any new wallpapers first and will show them first, it will then check to see if there are any older than a set number of days (this can be changed by the user), the hope for this is to prevent seeing the same wallpaper too often but to always get new wallpapers into rotation. Finally if both of the above return no result then a wallpaper will just be chosen at random. 

2. I have added user changeable options at the top. Now the user can change the location of folders A and B, and they can set the Age of wallpapers (in days) before seeing the same wallpaper again.